Objects don't move through screen boundaries

+++ include/vector2D.h
@@ -6,15 +6,15 @@
 class Vector2D {
     public:
         Vector2D () {};
-        Vector2D (u16 X, u16 Y);
+        Vector2D (s32 X, s32 Y);
         //virtual ~Vector2D();
         Vector2D operator + (Vector2D);
         Vector2D& operator += (Vector2D);
 
         void clear();
 
-        u16 x;
-        u16 y;
+        s32 x;
+        s32 y;
     private:
 };
